feat(EMS-4117-4118): update e2e test
